bug fixing

+++ aco_vrp_src/source/acsvrp/AntColony.java
@@ -69,10 +69,10 @@ public class AntColony {
 		try {
 			logger.trace("Loading AC parameters");
 		    properties.load(new FileInputStream("acsvrp.properties"));
-		    START_PHEROMON = Float.valueOf(properties.getProperty("START_PHEROMON"));
+		    START_PHEROMON = Double.valueOf(properties.getProperty("START_PHEROMON"));
 		    BETA = Float.valueOf(properties.getProperty("BETA"));
 		    ALPHA = Float.valueOf(properties.getProperty("ALPHA"));
-		    RO = Float.valueOf( properties.getProperty("RO"));
+		    RO = Double.valueOf( properties.getProperty("RO"));
 		    MAX_CYCLES_PARAM = Integer.valueOf( properties.getProperty("MAX_CYCLES_PARAM"));
 		    LOCAL_UPDATE = Boolean.valueOf( properties.getProperty("LOCAL_UPDATE"));
 		    SPEED = Integer.valueOf( properties.getProperty("SPEED"));
